fix: product service and controller

+++ src/main/java/br/com/dh/meli/projeto_integrador/repository/IShoppingCartRepository.java
@@ -4,5 +4,5 @@ import br.com.dh.meli.projeto_integrador.dto.ShoppingCartDTO;
 import br.com.dh.meli.projeto_integrador.model.ShoppingCart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IShoppingCartRepository extends JpaRepository<ShoppingCart,Integer> {
+public interface IShoppingCartRepository extends JpaRepository<ShoppingCart,Long> {
 }
